+++ src/main/java/kr/hhplus/be/server/order/infra/gateway/custom/OrderRepositoryCustom.java
@@ -1,9 +1,9 @@
-package kr.hhplus.be.server.order.infra.gateway;
+package kr.hhplus.be.server.order.infra.gateway.custom;
 
 import kr.hhplus.be.server.order.domain.Order;
 import kr.hhplus.be.server.order.domain.OutboxEvent;
 
 public interface OrderRepositoryCustom {
+
     Order save(Order order, OutboxEvent event);
-    OutboxEvent findByOrderId(Long orderId);
 }
