+++ src/main/java/kr/hhplus/be/server/order/infra/gateway/custom/OutboxRepositoryCustom.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.order.infra.gateway.custom;
+
+import kr.hhplus.be.server.order.domain.OutboxEvent;
+
+public interface OutboxRepositoryCustom {
+    OutboxEvent findByOrderId(Long orderId);
+}
